deleted CreateComment.tsx as it is no longer needed
